Improve module augmentation for `indent` plugin
