+++ packages/ckeditor5-indent/src/augmentation.ts
@@ -1,0 +1,35 @@
+/**
+ * @license Copyright (c) 2003-2023, CKSource Holding sp. z o.o. All rights reserved.
+ * For licensing, see LICENSE.md or https://ckeditor.com/legal/ckeditor-oss-license
+ */
+
+import type {
+	IndentBlockConfig,
+	Indent,
+	IndentBlock,
+	IndentUI,
+	IndentBlockCommand
+} from './index';
+
+declare module '@ckeditor/ckeditor5-core' {
+	interface EditorConfig {
+
+		/**
+		 * The configuration of the {@link module:indent/indentblock~IndentBlock block indentation feature}.
+		 *
+		 * Read more in {@link module:indent/indentconfig~IndentBlockConfig}.
+		 */
+		indentBlock?: IndentBlockConfig;
+	}
+
+	interface PluginsMap {
+		[ Indent.pluginName ]: Indent;
+		[ IndentBlock.pluginName ]: IndentBlock;
+		[ IndentUI.pluginName ]: IndentUI;
+	}
+
+	interface CommandsMap {
+		indentBlock: IndentBlockCommand;
+		outdentBlock: IndentBlockCommand;
+	}
+}
