+++ packages/ckeditor5-indent/src/indentblock.ts
@@ -15,8 +15,6 @@ import IndentUsingOffset from './indentcommandbehavior/indentusingoffset';
 import IndentUsingClasses from './indentcommandbehavior/indentusingclasses';
 import type { HeadingOption } from '@ckeditor/ckeditor5-heading';
 
-import './indentconfig';
-
 const DEFAULT_ELEMENTS = [ 'paragraph', 'heading1', 'heading2', 'heading3', 'heading4', 'heading5', 'heading6' ];
 
 /**
@@ -168,11 +166,5 @@ export default class IndentBlock extends Plugin {
 		}
 
 		this.editor.conversion.attributeToAttribute( definition );
-	}
-}
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ IndentBlock.pluginName ]: IndentBlock;
 	}
 }
